Visual part of end of year screen (#33)
* Visual part of end of year screen

* Fix analyzer errors

+++ lib/src/data/account.dart
@@ -4,5 +4,5 @@ part 'account.freezed.dart';
 
 @freezed
 class Account with _$Account {
-  const factory Account(int sum) = _Account;
+  const factory Account(double sum) = _Account;
 }
